+++ libexp/include/Galois/Runtime/hGraph.h
@@ -800,23 +800,23 @@ public:
       m.nedges = edge_counter;
 
       // copy memoization meta-data
-      m.num_master_nodes = (size_t *) calloc(hostNodes.size(), sizeof(size_t));;
-      m.master_nodes = (size_t **) calloc(hostNodes.size(), sizeof(size_t *));;
+      m.num_master_nodes = (unsigned int *) calloc(hostNodes.size(), sizeof(unsigned int));;
+      m.master_nodes = (unsigned int **) calloc(hostNodes.size(), sizeof(unsigned int *));;
       for(uint32_t h = 0; h < hostNodes.size(); ++h){
         m.num_master_nodes[h] = masterNodes[h].size();
         if (masterNodes[h].size() > 0) {
-          m.master_nodes[h] = (size_t *) calloc(masterNodes[h].size(), sizeof(size_t));;
+          m.master_nodes[h] = (unsigned int *) calloc(masterNodes[h].size(), sizeof(unsigned int));;
           std::copy(masterNodes[h].begin(), masterNodes[h].end(), m.master_nodes[h]);
         } else {
           m.master_nodes[h] = NULL;
         }
       }
-      m.num_slave_nodes = (size_t *) calloc(hostNodes.size(), sizeof(size_t));;
-      m.slave_nodes = (size_t **) calloc(hostNodes.size(), sizeof(size_t *));;
+      m.num_slave_nodes = (unsigned int *) calloc(hostNodes.size(), sizeof(unsigned int));;
+      m.slave_nodes = (unsigned int **) calloc(hostNodes.size(), sizeof(unsigned int *));;
       for(uint32_t h = 0; h < hostNodes.size(); ++h){
         m.num_slave_nodes[h] = slaveNodes[h].size();
         if (slaveNodes[h].size() > 0) {
-          m.slave_nodes[h] = (size_t *) calloc(slaveNodes[h].size(), sizeof(size_t));;
+          m.slave_nodes[h] = (unsigned int *) calloc(slaveNodes[h].size(), sizeof(unsigned int));;
           std::copy(slaveNodes[h].begin(), slaveNodes[h].end(), m.slave_nodes[h]);
         } else {
           m.slave_nodes[h] = NULL;
